pyproject.toml
[tool.poetry]
name = "github-readme-generator"
version = "0.1.0"
description = "A tool to generate README and other content from GitHub repositories"
authors = ["Adel Zaalouk"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
requests = "^2.28.1"
PyGithub = "^1.58.0"
GitPython = "^3.1.30"
python-dotenv = "^1.0.0"
openai = "^0.27.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.0"
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
github-readme-generator = "github_readme_generator.cli:cli"
